Rename: parse.py => script.py
The code in this file doesn't just do parsing anymore.
